Remove PTR record jazz and run ipa-client-install before ipa-replica-install
We are no longer trying to install the replica at boot, so there is no
need to wait until the PTR record is in place.  By the time the
terraform apply completes, the PTR record is guaranteed to be in
place.

Also remove the loop that waits until the master is up and running.
Since we would be setting up the servers manually, the master will be
completely set up before you even log into the replica.  This allows
us to stop passing master_hostname as a variable.

Finally, we run ipa-client-install before running
ipa-replica-install.  For some reason ipa-replica-install does not
appear to pass the value associated with the --principal flag to
ipa-client-install, so this is necessary.